Add files via upload
This is a seller's dashboard that I have completed a Python Script for.  It works with the eBay Developer API and pulls information for my sellers account into Power Bi.  It is somewhat unique because most seem to be server hosted scripts.  eBay's API is one of the worst for clarity on documentation.  Though, there is a SDK, I chose to make my own script to be able to customize, the way I liked.